+++ muna/beta/metadata/__init__.py
@@ -3,6 +3,7 @@
 #   Copyright © 2026 NatML Inc. All Rights Reserved.
 #
 
+from ._ort import OnnxRuntimeExecutionProvider, OnnxRuntimeOptimizationLevel
 from ._torch import TorchExporter
 from .coreml import CoreMLInferenceMetadata
 from .executorch import ExecuTorchInferenceBackend, ExecuTorchInferenceMetadata
@@ -10,15 +11,11 @@ from .iree import IREEInferenceBackend, IREEInferenceMetadata
 from .litert import LiteRTInferenceMetadata
 from .llama import LlamaCppBackend, LlamaCppInferenceMetadata
 from .mlx import MLXInferenceMetadata, MLXInferenceSessionMetadata
-from .onnx import (
-    OnnxRuntimeInferenceMetadata, OnnxRuntimeExecutionProvider,
-    OnnxRuntimeOptimizationLevel
-)
-from .onnxruntime import OnnxRuntimeInferenceSessionMetadata
+from .onnxruntime import OnnxRuntimeInferenceMetadata, OnnxRuntimeInferenceSessionMetadata
 from .openvino import OpenVINOInferenceMetadata
 from .qnn import QnnInferenceBackend, QnnInferenceMetadata, QnnInferenceQuantization
 from .tensorrt import (
-    CudaArchitecture, TensorRTInferenceMetadata,
+    CudaArchitecture, TensorRTInferenceMetadata, TensorRTInferenceSessionMetadata,
     TensorRTHardwareCompatibility, TensorRTPrecision
 )
 from .tensorrt_rtx import TensorRTRTXInferenceMetadata